+++ Braintree/ApplePayCard.cs
@@ -13,6 +13,7 @@ namespace Braintree
         public string SourceDescription { get; protected set; }
         public bool? IsDefault { get; protected set; }
         public string ImageUrl { get; protected set; }
+        public string CustomerId { get; protected set; }
         public DateTime? CreatedAt { get; protected set; }
         public DateTime? UpdatedAt { get; protected set; }
         public Subscription[] Subscriptions { get; protected set; }
@@ -28,6 +29,7 @@ namespace Braintree
             SourceDescription = node.GetString("source-description");
             IsDefault = node.GetBoolean("default");
             ImageUrl = node.GetString("image-url");
+            CustomerId = node.GetString("customer-id");
 
             CreatedAt = node.GetDateTime("created-at");
             UpdatedAt = node.GetDateTime("updated-at");